cloudflare-ddns-az-fn/requirements.txt
argparse
requests
tld
configparser
typing
azure-functions